+++ cmd/frontend/types/types.go
@@ -167,24 +167,6 @@ type SiteActivityPeriod struct {
 	RegisteredUserCount  int32
 	AnonymousUserCount   int32
 	IntegrationUserCount int32
-	Stages               *Stages
-}
-
-// NOTE: DO NOT alter this struct without making a symmetric change
-// to the updatecheck handler. This struct is marshalled and sent to
-// BigQuery, which requires the input match its schema exactly.
-type Stages struct {
-	Manage    int32 `json:"mng"`
-	Plan      int32 `json:"plan"`
-	Code      int32 `json:"code"`
-	Review    int32 `json:"rev"`
-	Verify    int32 `json:"ver"`
-	Package   int32 `json:"pkg"`
-	Deploy    int32 `json:"depl"`
-	Configure int32 `json:"conf"`
-	Monitor   int32 `json:"mtr"`
-	Secure    int32 `json:"sec"`
-	Automate  int32 `json:"auto"`
 }
 
 // NOTE: DO NOT alter this struct without making a symmetric change