+++ CHANGELOG.md
@@ -52,6 +52,7 @@ All notable changes to Sourcegraph are documented in this file.
 - The unused GraphQL types `DiffSearchResult` and `DeploymentConfiguration` were removed.
 - The deprecated GraphQL mutation `updateAllMirrorRepositories`.
 - The deprecated GraphQL field `Site.noRepositoriesEnabled`.
+- Total counts of users by product area have been removed from pings.
 - Aggregate daily, weekly, and monthly latencies (in ms) of code intelligence events (e.g., hover tooltips) have been removed from pings.
 
 ## 3.20.1